Account creation, update now take into account hashing strategy.
- Also touched password verification

Refers #1497
