+++ restcomm/restcomm.commons/src/main/java/org/restcomm/connect/commons/security/SecurityUtils.java
@@ -1,0 +1,40 @@
+/*
+ * TeleStax, Open Source Cloud Communications
+ * Copyright 2011-2014, Telestax Inc and individual contributors
+ * by the @authors tag.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation; either version 3 of
+ * the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>
+ *
+ */
+
+package org.restcomm.connect.commons.security;
+
+import org.apache.commons.codec.binary.Base64;
+
+import java.util.UUID;
+
+/**
+ * @author otsakir@gmail.com - Orestis Tsakiridis
+ */
+public class SecurityUtils {
+    public static String buildBasicAuthHeader(String username, String password) {
+        return "Basic " + Base64.encodeBase64String((username + ":" + password).getBytes());
+    }
+
+    // TODO is this is random enough ? probably yes.
+    public static String generateAccountAuthToken() {
+        return UUID.randomUUID().toString().replace("-","");
+    }
+
+}
