+++ restcomm/restcomm.http/src/main/java/org/restcomm/connect/http/client/rcmlserver/RcmlserverApi.java
@@ -31,7 +31,7 @@ import org.apache.log4j.Logger;
 import org.restcomm.connect.commons.common.http.CustomHttpClientBuilder;
 import org.restcomm.connect.commons.configuration.sets.MainConfigurationSet;
 import org.restcomm.connect.commons.configuration.sets.RcmlserverConfigurationSet;
-import org.restcomm.connect.commons.util.SecurityUtils;
+import org.restcomm.connect.commons.security.SecurityUtils;
 import org.restcomm.connect.commons.util.UriUtils;
 import org.restcomm.connect.dao.entities.Account;
 import org.restcomm.connect.http.exceptions.RcmlserverNotifyError;
